Implement the dir listing cache

+++ src/psrc/common/filesystem.c
@@ -124,7 +124,7 @@ char* strrelpath(const char* s) {
 void sanfilename_cb(const char* s, char r, struct charbuf* cb) {
     char c;
     #if (PLATFLAGS & PLATFLAG_WINDOWSLIKE)
-    int b = cb->len;
+    unsigned long b = cb->len;
     #endif
     while ((c = *s)) {
         if (ispathsep(c)) {if (r) cb_add(cb, r);}
@@ -135,9 +135,10 @@ void sanfilename_cb(const char* s, char r, struct charbuf* cb) {
         else cb_add(cb, c);
         ++s;
     }
+    // TODO: confirm that checking for illegal names can be omitted on PLAT_NXDK
     #if (PLATFLAGS & PLATFLAG_WINDOWSLIKE)
     if (b == cb->len) return;
-    int i = b;
+    unsigned long i = b;
     while (i < cb->len) {
         if (cb->data[i] == ' ') cb->data[i] = '_';
         else break;
@@ -145,7 +146,7 @@ void sanfilename_cb(const char* s, char r, struct charbuf* cb) {
     }
     if (i == b) {
         char* d = cb->data + b;
-        int l = cb->len - b;
+        unsigned long l = cb->len - b;
         if (((l == 3 || (l > 3 && d[3] == '.')) &&
              (!strncasecmp(d, "con", 3) || !strncasecmp(d, "prn", 3) || !strncasecmp(d, "aux", 3) || !strncasecmp(d, "nul", 3))) ||
             ((l == 4 || (l > 4 && d[4] == '.')) &&
